+++ frontend/src/App.tsx
@@ -23,7 +23,7 @@ import {
 import { CommandPalette } from './components/CommandPalette';
 import { SetupScreen } from './components/SetupScreen';
 import { Toaster } from './components/ui/sonner';
-import { WatchtowerNotifier } from './components/Watchtower/WatchtowerNotifier';
+import { WatchtowerAnnouncer } from './components/Watchtower/WatchtowerAnnouncer';
 import { useAppStore } from './lib/store';
 import { fetchModels, fetchServerInfo, fetchSavings, submitSavings, isTauri } from './lib/api';
 import { resolveModelSelection } from './lib/models';
@@ -285,7 +285,7 @@ export default function App() {
   return (
     <>
       <AppRoutes />
-      <WatchtowerNotifier />
+      <WatchtowerAnnouncer />
       <Toaster position="bottom-right" />
       {commandPaletteOpen && <CommandPalette />}
       {optInModalOpen && (